backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
redis==5.0.1
hashids==1.3.1
python-dotenv==1.0.0
gunicorn==21.2.0
